+++ src/components/Breadcrumbs/Breadcrumbs.tsx
@@ -170,7 +170,7 @@ export const Breadcrumbs: ChakraComponent<
       );
     }
 
-    const styles = useStyleConfig("CustomBreadcrumb", {
+    const styles = useStyleConfig("ReservoirBreadcrumb", {
       variant: breadcrumbsType,
     });
     const breadcrumbItems = getElementsFromData({