+++ src/components/Button/buttonChangelogData.ts
@@ -12,16 +12,19 @@ export const changelogData: ChangelogData[] = [
   {
     date: "Prerelease",
     version: "Prerelease",
+    type: "Update",
+    affects: ["Functionality", "Styles"],
+    notes: [
+      "Removes deprecated `link` buttonType.",
+      "Changes theme name from `CustomButton` to `ReservoirButton` for consistency.",
+    ],
+  },
+  {
+    date: "2025-05-22",
+    version: "3.6.2",
     type: "Bug Fix",
     affects: ["Styles"],
     notes: ["Fixes the style transitions."],
-  },
-  {
-    date: "Prerelease",
-    version: "Prerelease",
-    type: "Update",
-    affects: ["Functionality"],
-    notes: ["Removes deprecated `link` buttonType."],
   },
   {
     date: "2025-04-10",