+++ src/components/MultiSelectGroup/multiSelectGroupChangelogData.ts
@@ -14,6 +14,15 @@ export const changelogData: ChangelogData[] = [
     version: "Prerelease",
     type: "Update",
     affects: ["Styles"],
+    notes: [
+      "Removes use of `useNYPLBreakpoints` and replaces with equivalent CSS.",
+    ],
+  },
+  {
+    date: "2025-05-22",
+    version: "3.6.2",
+    type: "Update",
+    affects: ["Styles"],
     notes: ["Uses CSS to decrease reliance on `useNYPLBreakpoints` hook."],
   },
   {