+++ src/components/SearchBar/searchBarChangelogData.ts
@@ -13,6 +13,13 @@ export const changelogData: ChangelogData[] = [
     date: "Prerelease",
     version: "Prerelease",
     type: "Update",
+    affects: ["Accessibility", "Documentation"],
+    notes: ["Removes role=search on the component wrapper."],
+  },
+  {
+    date: "2025-05-22",
+    version: "3.6.2",
+    type: "Update",
     affects: ["Styles"],
     notes: [
       "Removes use of `useNYPLBreakpoints` and replaces with equivalent CSS.",