+++ src/components/Slider/Slider.tsx
@@ -10,7 +10,6 @@ import {
   SliderFilledTrack as ChakraSliderFilledTrack,
   SliderThumb as ChakraSliderThumb,
   SliderTrack as ChakraSliderTrack,
-  useColorMode,
   useMultiStyleConfig,
 } from "@chakra-ui/react";
 import React, { forwardRef } from "react";
@@ -133,7 +132,6 @@ export const Slider: ChakraComponent<
     }
 
     const [currentValue, setCurrentValue] = useStateWithDependencies(value);
-    const isDarkMode = useColorMode().colorMode === "dark";
 
     let finalIsInvalid = isInvalid;
     // In the Range Slider, if the first value is bigger than the second value,
@@ -150,8 +148,7 @@ export const Slider: ChakraComponent<
         "NYPL Reservoir Slider: The `min` prop is greater than the `max` prop."
       );
     }
-    const styles = useMultiStyleConfig("CustomSlider", {
-      isDarkMode,
+    const styles = useMultiStyleConfig("ReservoirSlider", {
       isDisabled,
       isInvalid: finalIsInvalid,
       showBoxes,