+++ CHANGELOG.md
@@ -21,6 +21,10 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 - Updates the Storybook font styles.
 - Updates the docs for the `Template` component.
 - Replaces the error thrown with a console.warn for the `Breadcrumbs`, `Heading`, and `Image` components.
+- Removes props with HTML equivalents and instructs devs to use the native attributes
+- Updates the grid layout of the content for the `Hero` `"primary"` variant.
+- Changes the names of theme objects that conflict with Chakra's base names from `Custom` to `Reservoir` for consistency. This impacts theme objects for `Breadcrumbs`, `Button`, `Select`, `Slider`, and `Table`.
+- Replaces positional function args with object references in utility functions/components when there are 3 or more arguments.
 
 ### Removals
 
@@ -37,18 +41,25 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
   - `StatusBadge`'s "low", "medium", "high" values from `statusBadgeTypeArray`
   - `StatusBadge`'s `level` prop
   - `StyledList`'s "tag", "mini" values in `textSizesArray`
+- Removes `role="search"` from `Searchbar` wrapper.
+- Removes use of `useNYPLBreakpoints` hook in `FilterBarInline` and `MultiSelectGroup`.
+- Removes `isDarkMode` function from `Slider` and `Accordion`.
 
 ## Prerelease
 
+## 3.6.2 (May 22, 2025)
+
 ### Adds
 
-- Adds "ARIA Landmarks" accessibility guide
+- Adds the `ARIA Landmarks` page to the `Accessibility Guide` in Storybook.
+- Adds explicit `tabs` design tokens for font size and weight.
+- Adds the `Progressive Enhancement` page to the `Accessibility Guide` in Storybook.
 
 ### Updates
 
-- Update Storybook and related npm packages to `v8.6.12`.
+- Updates Storybook and related npm packages to `v8.6.12`.
 - Updates `vite` to `v5.4.18`, `@vitejs/plugin-react` to `v4.4.1`, and `vite-plugin-svgr` to `v4.3.0`.
-- Replaces positional function args with object references in utility functions/components when there are 3 or more arguments.
+- Updates the `Tabs` component to use the explicitly defined `tabs` design tokens for font size and weight.
 
 ### Fixes
 
@@ -56,7 +67,7 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 ### Removes
 
-- Removes use of `useNYPLBreakpoints` hook where possible without causing breaking change.
+- Removes use of `useNYPLBreakpoints` hook where possible without causing breaking change. This update impacts the `FeedbackBox`, `Modal`, `MultiSelectGroup`, `NewsletterSignup`, `SearchBar`, and `Table` components.
 
 ## 3.6.1 (April 24, 2025)